Personal fix for issue: #2566
When an HTML form is used to submit data to DRF this function parse_html_list is called.
This functions can't parse a list of strings.

To work around this, post the data as a valid JSON list: 

```["example string 1", "test string 2"]```

This list is parsed as JSON and returned. Thus allowing a formdata with a list to be submitted.